rewrite most tests to jest
