--- test/integration/test-to-string.js
@@ -1,5 +1,0 @@
-var common   = require('../common');
-var assert   = common.assert;
-var FormData = require(common.dir.lib + '/form_data');
-
-assert(new FormData().toString() === '[object FormData]');
