+++ test/old.test.js
@@ -8,21 +8,21 @@ let httpServer;
 
 beforeAll(async () => {
   // start static server for all tests
-  httpServer = await new Promise((resolve) => staticServer(resolve));
+  httpServer = await staticServer();
 })
 
 afterAll(async () => {
   await new Promise((resolve) => httpServer.close(resolve));
 })
 
-it('should run all tests', async () => {
-  // todo rewrite tests to jest
-  const testFilesDir = join(__dirname, 'integration');
+// todo rewrite tests to jest
+it('should run old tests', async () => {
+  const testFilesDir = join(__dirname, 'old');
   const testFiles = await fs.readdir(testFilesDir);
 
   for (const testFile of testFiles) {
     const testFilePath = join(testFilesDir, testFile);
-    console.log('running', testFile)
+    console.log('running', testFile);
     await new Promise((resolve, reject) => {
       const cp = childProcess.fork(testFilePath);
       // https://github.com/facebook/jest/issues/3190
