+++ test/options-override.test.js
@@ -1,0 +1,8 @@
+var common = require('./common');
+var assert = common.assert;
+var FormData = require(common.dir.lib + '/form_data');
+
+it('options-override', () => {
+  var form = new FormData({maxDataSize: 20 * 1024 * 1024});
+  assert.strictEqual(form.maxDataSize, 20 * 1024 * 1024);
+});
